# intelligent_reimbursement_system_server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/reimbursement-types": {
      "get": {
        "operationId": "ReimbursementTypeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取报销类型列表",
        "tags": [
          "报销类型"
        ]
      },
      "post": {
        "operationId": "ReimbursementTypeController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "创建报销类型",
        "tags": [
          "报销类型"
        ]
      }
    },
    "/api/reimbursement-types/{id}": {
      "put": {
        "operationId": "ReimbursementTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReimbursementTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "更新报销类型",
        "tags": [
          "报销类型"
        ]
      },
      "delete": {
        "operationId": "ReimbursementTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "删除报销类型",
        "tags": [
          "报销类型"
        ]
      }
    },
...
    "/api/v1/opinions": {
      "post": {
        "operationId": "OpinionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpinionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "提交意见反馈",
        "tags": [
          "意见反馈"
        ]
      },
      "get": {
        "operationId": "OpinionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有意见反馈（管理员）",
        "tags": [
          "意见反馈"
        ]
      }
    },
    "/api/v1/opinions/mine": {
      "get": {
        "operationId": "OpinionController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取我的意见反馈",
        "tags": [
          "意见反馈"
        ]
      }
    },
    "/api/v1/opinions/{id}/status": {
      "patch": {
        "operationId": "OpinionController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOpinionStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改反馈状态",
        "tags": [
          "意见反馈"
        ]
      }
    },
...
    "/api/employees": {
      "get": {
        "operationId": "EmployeeController_findAll",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dept_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取员工列表",
        "tags": [
          "员工"
        ]
      },
      "post": {
        "operationId": "EmployeeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建员工",
        "tags": [
          "员工"
        ]
      }
    },
    "/api/employees/{id}": {
      "put": {
        "operationId": "EmployeeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新员工",
        "tags": [
          "员工"
        ]
      },
      "delete": {
        "operationId": "EmployeeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除员工",
        "tags": [
          "员工"
        ]
      }
    },
...
    "/api/menus": {
      "get": {
        "operationId": "MenuController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取菜单树",
        "tags": [
          "菜单"
        ]
      },
      "post": {
        "operationId": "MenuController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建菜单",
        "tags": [
          "菜单"
        ]
      }
    },
    "/api/menus/flat": {
      "get": {
        "operationId": "MenuController_findAllFlat",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取菜单列表（平铺）",
        "tags": [
          "菜单"
        ]
      }
    },
    "/api/menus/{id}": {
      "get": {
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取单个菜单",
        "tags": [
          "菜单"
        ]
      },
      "put": {
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新菜单",
        "tags": [
          "菜单"
        ]
      },
      "delete": {
        "operationId": "MenuController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除菜单",
        "tags": [
          "菜单"
        ]
      }
    }
  },
  "info": {
    "title": "智能报销系统 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateReimbursementTypeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "export_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "formula": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "over_limit_threshold": {
            "type": "number"
          }
        }
      },
...
      "CreateOpinionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "标题"
          },
          "content": {
            "type": "string",
            "description": "意见内容"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "UpdateOpinionStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "状态：0待处理 1已处理 2已关闭",
            "enum": [
              0,
              1,
              2
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "employee_no": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "type": "number"
          },
          "dept_id": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "status": {
            "type": "number"
          }
        },
        "required": [
          "employee_no",
          "name",
          "gender"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "employee_no": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "type": "number"
          },
          "dept_id": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "status": {
            "type": "number"
          }
        }
      },
...
      "CreateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "sort": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "directory",
              "menu",
              "button"
            ]
          },
          "parent_id": {
            "type": "string"
          },
          "permission": {
            "type": "string"
          },
          "visible": {
            "type": "number"
          },
          "status": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "sort": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "directory",
              "menu",
              "button"
            ]
          },
          "parent_id": {
            "type": "string"
          },
          "permission": {
            "type": "string"
          },
          "visible": {
            "type": "number"
          },
          "status": {
            "type": "number"
          }
        }
      }
    }
  }
}